+++ src/Frontend/Pages/Auth/index.css
@@ -41,7 +41,7 @@
 }
 .auth-left > .auth-left-desc {
   padding: 0.5rem 0.5rem;
-  color: rgb(216, 216, 216);
+  color: var(--app-pale-white);
 }
 .auth-left > img {
   width: 80%;
@@ -57,7 +57,7 @@
 .auth-email > input {
   border: none;
   outline: none;
-  border-bottom: 1px solid #e0e0e0;
+  border-bottom: 1px solid var(--app-light-white);
   width: 100%;
 }
 
@@ -131,12 +131,12 @@
 }
 
 .has-float-label input:focus {
-  border-bottom: 2px solid #2979ff;
+  border-bottom: 2px solid var(--primary);
   transition: all 0.2s;
 }
 
 .has-float-label input:disabled {
-  border-bottom: 2px dotted #8a8a8a !important;
+  border-bottom: 2px dotted var(--app-dark-white);
   pointer-events: none;
   background-color: inherit;
 }
@@ -148,8 +148,8 @@
   border-radius: 0;
   box-shadow: none;
   background-color: initial;
-  caret-color: #2979ff;
-  border-bottom: 2px solid #e0e0e0;
+  caret-color: var(--primary);
+  border-bottom: 2px solid var(--app-light-white);
   width: 100%;
 }
 
@@ -165,7 +165,7 @@ input {
   font-size: 14px;
   top: 20px;
   bottom: -8px;
-  color: #9c9999;
+  color: var(--app-pale-gray);
 }
 
 .has-float-label label,
@@ -175,18 +175,18 @@ input {
   top: 0;
   cursor: text;
   font-size: 12px;
-  color: #6d6d6d;
+  color: var(--app-deep-gray);
   transition: all 0.2s;
   pointer-events: none;
 }
 
 .has-float-label input:focus ~ .label {
-  color: #2979ff;
+  color: var(--primary);
 }
 
 .helper {
   display: block;
-  color: #6d6d6d;
+  color: var(--app-deep-gray);
 }
 
 .helper,
@@ -201,5 +201,5 @@ a {
 }
 
 a:visited .helper {
-  color: #eee;
+  color: var(--app-pale-white);
 }
