+++ src/Frontend/Pages/Home/Category/index.css
@@ -6,8 +6,3 @@
   gap: 2rem;
   box-shadow: 0 1px 1px 0 rgb(0 0 0 / 16%);
 }
-
-/* .category > img {
-  border-style: solid;
-  border-color: red;
-} */
