+++ src/Frontend/Pages/Home/Featured/index.css
@@ -3,7 +3,7 @@
   margin: 0.5rem;
 }
 .featured-header > .kash-h5 {
-  font-weight: 800 !important;
+  font-weight: 800;
   text-align: left;
   margin: 0 0 0 0.3rem;
 }
